fix: update references to `SplitDiffView` after module rename
