+++ demo/diff/index.html
@@ -65,7 +65,7 @@
     });
 
     require([
-        "ace/ext/diff/diff_view",
+        "ace/ext/diff/split_diff_view",
         "ace/ext/diff/inline_diff_view",
         "demo/kitchen-sink/util",
         "demo/kitchen-sink/layout", 
@@ -74,7 +74,7 @@
         "ace/ext/menu_tools/overlay_page",
     ], function () {
         var {InlineDiffView} = require("ace/ext/diff/inline_diff_view");
-        var {SplitDiffView} = require("ace/ext/diff/diff_view");
+        var {SplitDiffView} = require("ace/ext/diff/split_diff_view");
         var {DiffProvider} = require("ace/ext/diff/providers/default");
         var {OptionPanel, optionGroups} = require("ace/ext/options")
         var {buildDom} = require("ace/lib/dom");
