+++ src/ext/diff/split_diff_view.js
